Created the basic search system
The search system essentually wraps around SQL's SELECT statement. this
program, however, allows the user to search according to specific
criteria, or just leave the field blank to search the entire database.
Pressing 'q' will exit during a database listing, but it does this by
returning an error. Hopefully, this will be remedied by the time I move to
the GUI.

There are four main "operations": search, new, update and delete, each of
which uses operations.h as it's access point. operations.cpp will
eventually be deleted. I've also created a dirty struct called Counter,
that can be used to keep track of calls to callbacks, so I can print out
the table headers, and pause the listing once per page.

Finally, I've had to adjust console_colors to accomodate this program.
I'm considering moving this to the Codebase, since it's pretty useful.